requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
joblib
pydantic
requests
python-dotenv
torch
matplotlib